Rename views.py to route_api.py
