+++ orlo/route_import.py
@@ -4,7 +4,7 @@ import json
 from flask import jsonify, request
 from orlo import app
 from orlo.orm import db, Package, Release, PackageResult, ReleaseNote, Platform
-from orlo.views import validate_request_json
+from orlo.util import validate_request_json
 from sqlalchemy.orm import exc
